refactor(usage-logs): own usage log model types
