+++ e2e/alarmSchedulers.spec.ts
@@ -3,7 +3,6 @@ import type { BrowserContext, Page, Route, Worker } from "@playwright/test"
 import { ChannelType } from "~/constants"
 import { OPTIONS_PAGE_PATH } from "~/constants/extensionPages"
 import { SITE_TYPES } from "~/constants/siteType"
-import { LogType } from "~/services/apiService/common/type"
 import { STORAGE_KEYS } from "~/services/core/storageKeys"
 import { DAILY_BALANCE_HISTORY_ALARM_NAME } from "~/services/history/dailyBalanceHistory/constants"
 import { getDayKeyFromUnixSeconds as getBalanceHistoryDayKeyFromUnixSeconds } from "~/services/history/dailyBalanceHistory/dayKeys"
@@ -22,6 +21,7 @@ import {
   WebdavAutoSyncMessageTypes,
 } from "~/services/runtimeMessaging/messageTypes"
 import { SITE_ANNOUNCEMENTS_ALARM_NAME } from "~/services/siteAnnouncements/constants"
+import { LogType } from "~/services/usageLogs/model"
 import { AUTO_CHECKIN_SCHEDULE_MODE } from "~/types/autoCheckin"
 import type { AutoCheckinStatus } from "~/types/autoCheckin"
 import type { DailyBalanceHistoryStore } from "~/types/dailyBalanceHistory"
