+++ tests/services/usageHistory/core.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from "vitest"
 
-import { LogType, type LogItem } from "~/services/apiService/common/type"
 import { USAGE_HISTORY_LIMITS } from "~/services/history/usageHistory/constants"
 import {
   computeRetentionCutoffDayKey,
@@ -11,6 +10,8 @@ import {
   USAGE_HISTORY_LATENCY_BUCKET_UPPER_BOUNDS_SECONDS,
   USAGE_HISTORY_SLOW_THRESHOLD_SECONDS,
 } from "~/services/history/usageHistory/core"
+import { LogType } from "~/services/usageLogs/model"
+import type { LogItem } from "~/services/usageLogs/model"
 
 /**
  * Create a fully populated Consume log item for usage-history unit tests.
